+++ CHANGELOG.md
@@ -14,6 +14,7 @@ size to the area for a single id card. If more than one foodsaver is selected, t
 and an A4 sheet is used. !2198 @chriswalg
 - Send a bell notification and email, if a foodsaver was verified. !2198 @chriswalg
 - Set fileformat jpg / png for wallposts !2229 @chriswalg
+- Removed the possibility to edit the member list on a working group's edit page !2228 @alex.simm
 
 ## Bugfixes
 - When going back during registration, email and password are unchanged. !1114 @AntonBallmaier